+++ _studio/shared/umc/codec/h265_dec/src/umc_h265_widevine_supplier.cpp
@@ -173,7 +173,7 @@ UMC::Status WidevineTaskSupplier::ParseWidevineVPSSPSPPS(DecryptParametersWrappe
             if (!sps.getPTL()->GetGeneralPTL()->level_idc && sps.sps_max_dec_pic_buffering[HighestTid])
             {
                 Ipp32u level_idc = levelIndexArray[0];
-                for (int i = 0; i < sizeof(levelIndexArray)/sizeof(levelIndexArray[0]); i++)
+                for (size_t i = 0; i < sizeof(levelIndexArray)/sizeof(levelIndexArray[0]); i++)
                 {
                     level_idc = levelIndexArray[i];
                     newDPBsize = (Ipp8u)CalculateDPBSize(level_idc,
